1494 - Move functions to utils
